requirements.txt
aniso8601==1.1.0
BeautifulSoup==3.2.1
click==6.6
cssselect==0.9.2
Flask==0.11.1
Flask-RESTful==0.3.5
goose-extractor==1.0.25
itsdangerous==0.24
jieba==0.38
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
nltk==3.2.1
Pillow==3.4.1
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
Werkzeug==0.11.11
git+https://github.com/mit-nlp/MITIE.git
geopy==1.11.0
goose-extractor==1.0.25
